+++ deode/templates/stand_alone.py
@@ -1,6 +1,6 @@
 """NoSchedulerTemplate."""
 
-# @ENV_SUB@
+import os
 
 from deode.config_parser import ConfigParserDefaults, ParsedConfig
 from deode.derived_variables import derived_variables, set_times
@@ -36,7 +36,7 @@ def default_main(task, config, deode_home):
 
 
 if __name__ == "__main__":
-    TASK_NAME = "@STAND_ALONE_TASK_NAME@"
-    CONFIG = "@STAND_ALONE_TASK_CONFIG@"
-    DEODE_HOME = "@STAND_ALONE_DEODE_HOME@"
+    TASK_NAME = os.environ["STAND_ALONE_TASK_NAME"]
+    CONFIG = os.environ["STAND_ALONE_TASK_CONFIG"]
+    DEODE_HOME = os.environ["STAND_ALONE_DEODE_HOME"]
     default_main(TASK_NAME, CONFIG, DEODE_HOME)